// src/models/resistencia.model.ts
export interface Resistencia {
    tipo: string; // Forgeign Key UUID
    resistencia_1: string; // Foreign Key UUID referencing tipos
    resistencia_2: string; // Foreign Key UUID referencing tipos
    resistencia_3: string; // Foreign Key UUID referencing tipos
    resistencia_4: string; // Foreign Key UUID referencing tipos
    resistencia_5: string; // Foreign Key UUID referencing tipos
    resistencia_6: string; // Foreign Key UUID referencing tipos
    resistencia_7: string; // Foreign Key UUID referencing tipos
    resistencia_8: string; // Foreign Key UUID referencing tipos
    resistencia_9: string; // Foreign Key UUID referencing tipos
    resistencia_10: string; // Foreign Key UUID referencing tipos
    key_resistencias: string; // Primary Key UUID
}